refactor(discord): rebase gate shard onto current main
Reapply the verified R4 extraction unchanged because the Discord adapter has not changed between d762ed9b and f43eabee.

Part of #78647
Part of #78634
Part of #79564
Part of #79772
